refactor: structured coupon folder
